+++ src/Executor/ExecutionResult.php
@@ -19,20 +19,25 @@ use Throwable;
  * Could be converted to [spec-compliant](https://facebook.github.io/graphql/#sec-Response-Format)
  * serializable array using `toArray()`.
  *
+ * @see Throwable
+ *
  * @phpstan-type SerializableError array{
  *   message: string,
  *   locations?: array<int, array{line: int, column: int}>,
  *   path?: array<int, int|string>,
- *   extensions?: array<string, mixed>,
+ *   extensions?: array<string, mixed>
  * }
  * @phpstan-type SerializableErrors array<int, SerializableError>
  * @phpstan-type SerializableResult array{
  *     data?: array<string, mixed>,
  *     errors?: SerializableErrors,
- *     extensions?: array<string, mixed>,
+ *     extensions?: array<string, mixed>
  * }
  * @phpstan-type ErrorFormatter callable(Throwable): SerializableError
  * @phpstan-type ErrorsHandler callable(array<Error> $errors, ErrorFormatter $formatter): SerializableErrors
+ *
+ * @see https://github.com/vimeo/psalm/issues/6928
+ * @psalm-type ErrorsHandler callable(Error[], ErrorFormatter): SerializableErrors
  */
 class ExecutionResult implements JsonSerializable
 {