+++ src/Executor/Promise/Adapter/AmpPromiseAdapter.php
@@ -8,8 +8,10 @@ use function Amp\Promise\all;
 use Amp\Promise as AmpPromise;
 use Amp\Success;
 use function array_replace;
+use function assert;
 use GraphQL\Executor\Promise\Promise;
 use GraphQL\Executor\Promise\PromiseAdapter;
+use function is_array;
 use Throwable;
 
 class AmpPromiseAdapter implements PromiseAdapter
@@ -77,8 +79,13 @@ class AmpPromiseAdapter implements PromiseAdapter
         return new Promise($promise, $this);
     }
 
-    public function all(array $promisesOrValues): Promise
+    public function all(iterable $promisesOrValues): Promise
     {
+        assert(
+            is_array($promisesOrValues),
+            'AmpPromiseAdapter::all(): Argument #1 ($promisesOrValues) must be of type array'
+        );
+
         /** @var array<AmpPromise<mixed>> $promises */
         $promises = [];
         foreach ($promisesOrValues as $key => $item) {