+++ src/Language/AST/NodeList.php
@@ -28,10 +28,12 @@ class NodeList implements ArrayAccess, IteratorAggregate, Countable
     private $nodes;
 
     /**
-     * @param array<Node|array<string, mixed>> $nodes
-     * @phpstan-param array<T|array<string, mixed>> $nodes
+     * @template TT of Node
      *
-     * @phpstan-return self<T>
+     * @param array<Node|array<string, mixed>> $nodes
+     * @phpstan-param array<TT|array<string, mixed>> $nodes
+     *
+     * @phpstan-return self<TT>
      */
     public static function create(array $nodes): self
     {