+++ src/Type/Definition/InputObjectType.php
@@ -21,7 +21,7 @@ use function is_string;
  *   description?: string|null,
  *   fields: iterable<FieldConfig>|callable(): iterable<FieldConfig>,
  *   astNode?: InputObjectTypeDefinitionNode|null,
- *   extensionASTNodes?: array<int, InputObjectTypeExtensionNode>|null,
+ *   extensionASTNodes?: array<int, InputObjectTypeExtensionNode>|null
  * }
  */
 class InputObjectType extends Type implements InputType, NullableType, NamedType
@@ -155,7 +155,6 @@ class InputObjectType extends Type implements InputType, NullableType, NamedType
             $fields = $fields();
         }
 
-        // @phpstan-ignore-next-line should not happen if used correctly
         if (! is_iterable($fields)) {
             $invalidFields = Utils::printSafe($fields);